app/requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
visitor==0.1.3
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0
